% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdata}
\alias{fdata}
\title{Frequency and Specimens Data}
\format{
An object of class \code{data.frame} with 3771 rows and 4 columns.
}
\usage{
fdata
}
\description{
Example Frequency and Specimens data by year for each species to be used in `PopulationGrowthR` package
}
\keyword{datasets}
