% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2freqdata.R
\name{raw2freqdata}
\alias{raw2freqdata}
\title{Extract Frequency and Specimen data from the raw data}
\usage{
raw2freqdata(rawdata, species = "species", year = "year")
}
\arguments{
\item{rawdata}{a dataframe containing species, year}

\item{species}{name of the column containing species names}

\item{year}{name of the column containing year}
}
\value{
Retirns a list of two dataframes
\itemize{
\item data - a dataframe conatining Species, Year, Freuqency and Specimens
\item yeardata - a dataframe containing Year and Specimens
}
}
\description{
Extract Frequency and Specimen data from the raw data
}
\examples{
cleandata = raw2freqdata(rawdata)
fdata = cleandata$data
yeardata = cleandata$yeardata
}
