% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Raw GBIF Data}
\format{
An object of class \code{data.frame} with 34088 rows and 50 columns.
}
\usage{
rawdata
}
\description{
Example raw GBIF data used in `PopulationGrowthR` package
}
\keyword{datasets}
