% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeardata}
\alias{yeardata}
\title{Total Specimens Data}
\format{
An object of class \code{data.frame} with 60 rows and 2 columns.
}
\usage{
yeardata
}
\description{
Example total Specimens data by year to be used in `PopulationGrowthR` package
}
\keyword{datasets}
